Move things out of orbital server and into launcher
