+++ filesystem/schemes/orbital/display/mod.rs
@@ -7,45 +7,9 @@ use orbital::Color;
 use orbital::Point;
 use orbital::Size;
 
-/// The info of the VBE mode
-#[derive(Copy, Clone)]
-#[repr(packed)]
-pub struct VBEModeInfo {
-    attributes: u16,
-    win_a: u8,
-    win_b: u8,
-    granularity: u16,
-    winsize: u16,
-    segment_a: u16,
-    segment_b: u16,
-    winfuncptr: u32,
-    bytesperscanline: u16,
-    xresolution: u16,
-    yresolution: u16,
-    xcharsize: u8,
-    ycharsize: u8,
-    numberofplanes: u8,
-    bitsperpixel: u8,
-    numberofbanks: u8,
-    memorymodel: u8,
-    banksize: u8,
-    numberofimagepages: u8,
-    unused: u8,
-    redmasksize: u8,
-    redfieldposition: u8,
-    greenmasksize: u8,
-    greenfieldposition: u8,
-    bluemasksize: u8,
-    bluefieldposition: u8,
-    rsvdmasksize: u8,
-    rsvdfieldposition: u8,
-    directcolormodeinfo: u8,
-    physbaseptr: u32,
-    offscreenmemoryoffset: u32,
-    offscreenmemsize: u16,
-}
+use self::vbe::VBEMODEINFO;
 
-const VBEMODEINFO: *const VBEModeInfo = 0x5200 as *const VBEModeInfo;
+mod vbe;
 
 /// A display
 pub struct Display {
